# openapi.yaml
openapi: 3.1.0
info:
  title: Archer API
  version: 0.1.0
  description: API for the Archer freelance marketplace. Amounts use integer minor units; Archer does not process payments.
servers:
  - url: http://localhost:4000/api/v1
...
paths:
...
  /client-profile:
    put:
      tags: [Profiles]
      security: [{ bearerAuth: [] }]
      summary: Create or replace the current client profile
      responses: { "200": { description: Client profile } }
  /freelancer-profile:
    put:
      tags: [Profiles]
      security: [{ bearerAuth: [] }]
      summary: Create or replace the current freelancer profile
      responses: { "200": { description: Freelancer profile } }
  /freelancers:
    get:
      tags: [Profiles]
      summary: Discover freelancers
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: skill, in: query, schema: { type: string } }
        - { name: currency, in: query, schema: { $ref: "#/components/schemas/Currency" } }
      responses: { "200": { description: Paginated freelancers } }
  /freelancers/{username}:
    get:
      tags: [Profiles]
      summary: Get a public freelancer profile
      parameters: [{ $ref: "#/components/parameters/Username" }]
      responses: { "200": { description: Freelancer profile }, "404": { $ref: "#/components/responses/NotFound" } }
...
  /jobs/{jobId}/proposals:
    get:
      tags: [Proposals]
      security: [{ bearerAuth: [] }]
      summary: List proposals for an owned job
      parameters: [{ $ref: "#/components/parameters/JobId" }]
      responses: { "200": { description: Proposals } }
    post:
      tags: [Proposals]
      security: [{ bearerAuth: [] }]
      summary: Submit a proposal
      parameters: [{ $ref: "#/components/parameters/JobId" }]
      responses: { "201": { description: Proposal submitted } }
  /proposals/{proposalId}/accept:
    post:
      tags: [Proposals]
      security: [{ bearerAuth: [] }]
      summary: Accept a proposal and create a contract offer
      parameters: [{ $ref: "#/components/parameters/ProposalId" }]
      responses: { "201": { description: Contract offered } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Username: { name: username, in: path, required: true, schema: { type: string } }
    JobId: { name: jobId, in: path, required: true, schema: { type: string } }
    ProposalId: { name: proposalId, in: path, required: true, schema: { type: string } }
...
  schemas:
    Currency: { type: string, enum: [USD, MMK] }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties: { code: { type: string }, message: { type: string }, fields: { type: object, additionalProperties: { type: array, items: { type: string } } }, requestId: { type: string } }
  responses:
...
    NotFound: { description: Resource not found, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...